Move util function to output.
